Sign-Language-main/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.7.0.68
mediapipe
scikit-learn>=1.3.0
numpy>=1.21.0
Pillow>=9.0.0

